requirements.txt
aiofiles==23.1.0
aiogram @ https://github.com/aiogram/aiogram/archive/refs/heads/dev-3.x.zip#sha256=c8bc350d5ab3cc59319a3b22e1a3f43043462f5f82a0f8aaf7f3a8ddaf49c80f
aiohttp==3.8.4
aioredis==2.0.1
aiosignal==1.3.1
appdirs==1.4.4
async-timeout==4.0.2
asyncpg==0.27.0
attrs==23.1.0
Babel==2.9.1
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
cssselect==1.2.0
exceptiongroup==1.1.1
fake-useragent==1.1.3
feedparser==6.0.10
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
importlib-metadata==6.5.0
importlib-resources==5.12.0
iniconfig==2.0.0
lxml==4.9.2
magic-filter==1.0.9
multidict==6.0.4
numpy==1.24.2
packaging==23.1
pandas==2.0.0
parse==1.19.0
pluggy==1.0.0
pydantic==1.10.7
pyee==8.2.2
pygismeteo==5.0.6
pygismeteo-base==4.1.0
pyppeteer==1.0.2
pyquery==2.0.0
pytest==7.3.1
pytest-asyncio==0.21.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
redis==4.5.4
requests==2.28.2
requests-html==0.10.0
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.9
tomli==2.0.1
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
w3lib==2.1.1
websockets==10.4
yahoo-fin==0.8.9.1
yarl==1.8.2
zipp==3.15.0
